@@ -1,24 +1,34 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
-import { combineReducers, createStore } from 'redux';
+// import { Provider } from 'react-redux';
+// import { combineReducers, createStore } from 'redux';
+// import productReducer from './store/reducers/products';
 import { BrowserRouter } from 'react-router-dom';
-
+import ProductsProvider from './context/products-context';
 import './index.css';
 import App from './App';
-import productReducer from './store/reducers/products';
 
-const rootReducer = combineReducers({
+/* const rootReducer = combineReducers({
   shop: productReducer
-});
+}); 
+//
+const store = createStore(rootReducer); */
 
-const store = createStore(rootReducer);
-
-ReactDOM.render(
-  <Provider store={store}>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
-  </Provider>,
-  document.getElementById('root')
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(<ProductsProvider>
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>
+</ProductsProvider>
 );
+
+// ReactDOM.render(
+//   // <Provider store={store}>
+//   <ProductsProvider>
+//     <BrowserRouter>
+//       <App />
+//     </BrowserRouter>
+//   </ProductsProvider>,
+//   // </Provider>,
+//   document.getElementById('root')
+// );
