@@ -1,11 +1,11 @@
-import React from 'react';
-import { useSelector } from 'react-redux';
-
+import React, { useContext } from 'react';
+// import { useSelector } from 'react-redux';
+import ProductsContext from '../context/products-context';
 import ProductItem from '../components/Products/ProductItem';
 import './Products.css';
 
 const Products = props => {
-  const productList = useSelector(state => state.shop.products);
+  const productList = useContext(ProductsContext).products;
   return (
     <ul className="products-list">
       {productList.map(prod => (
